move code around to downstream.rs
